+++ src/main/java/mil/nga/geopackage/tiles/TileGenerator.java
@@ -352,8 +352,9 @@ public abstract class TileGenerator {
 			// Create the srs if needed
 			SpatialReferenceSystemDao srsDao = geoPackage
 					.getSpatialReferenceSystemDao();
-			SpatialReferenceSystem srs = srsDao.getOrCreateCode(projection
-					.getAuthority(), Long.parseLong(projection.getCode()));
+			SpatialReferenceSystem srs = srsDao.getOrCreateCode(
+					projection.getAuthority(),
+					Long.parseLong(projection.getCode()));
 			// Create the tile table
 			tileMatrixSet = geoPackage.createTileTableWithMetadata(tableName,
 					tileGridBoundingBox, srs.getSrsId(), tileGridBoundingBox,
@@ -458,8 +459,8 @@ public abstract class TileGenerator {
 		// Set the tile matrix set bounding box to be the world
 		BoundingBox standardWgs84Box = new BoundingBox(
 				-ProjectionConstants.WGS84_HALF_WORLD_LON_WIDTH,
-				ProjectionConstants.WGS84_HALF_WORLD_LON_WIDTH,
 				ProjectionConstants.WEB_MERCATOR_MIN_LAT_RANGE,
+				ProjectionConstants.WGS84_HALF_WORLD_LON_WIDTH,
 				ProjectionConstants.WEB_MERCATOR_MAX_LAT_RANGE);
 		ProjectionTransform wgs84ToWebMercatorTransform = ProjectionFactory
 				.getProjection(ProjectionConstants.EPSG_WORLD_GEODETIC_SYSTEM)