+++ src/test/java/mil/nga/geopackage/test/tiles/UrlTileGeneratorCreateTest.java
@@ -14,6 +14,13 @@ import org.junit.Test;
  */
 public class UrlTileGeneratorCreateTest extends CreateGeoPackageTestCase {
 
+	@Override
+	public void geoPackageSetUp() throws Exception {
+		super.geoPackageSetUp();
+
+		UrlTileGeneratorUtils.checkUrl();
+	}
+
 	/**
 	 * Constructor
 	 */